Replace placeholder option with preserveContent
I wasn’t happy to expose raw, complicated regexes, this moves us to a simple boolean to control behaviour for a specific tag. The implementation is still lacking by placeholders being applied across all tags, because rendering doesn’t know which tag is active. Might be worth expanding in the future, but for now curious to see this providing any value at all. 